pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "kavalkilu"
version = '2.0.2'
description = "A library for integrating some common home automation components"
authors = ["bobrock <bobrock@tuta.io>"]
license = "MIT"
readme = 'README.md'
repository = 'https://github.com/barretobrock/kavalkilu'
packages = [
    { include = 'kavalkilu' },
]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = "^3.10"
pukr = { git = "https://github.com/barretobrock/pukr.git", branch = "main" }
influxdb = "~5"
numpy = "~1.22"
pandas = "~1.4"
paho-mqtt = "~1"
psutil = "~5"
python-dateutil = "~2"
requests = "~2"
# Optional dependencies would go down here
# example = { version = ">=1.7.0", optional = true }

[tool.poetry.dev-dependencies]
pre-commit = "~2"
pytest = "~7"
pytest-cov = "~3"
flake8 = "~4"
tox = "~3"

[tool.isort]
profile = 'black'
force_grid_wrap = 2
force_sort_within_sections = true
skip_gitignore = true
skip_glob = ['*_pb2.py', '*_pb2_grpc.py']

[tool.poetry.extras]
test = ["pytest"]
